Renamed old InterfacesNode file to better represent ComponentNode
