+++ src/dependency-manager/src/graph/index.ts
@@ -5,7 +5,7 @@ import OutputEdge from './edge/output';
 import ServiceEdge from './edge/service';
 import { DependencyNode } from './node';
 import GatewayNode from './node/gateway';
-import ComponentNode from './node/interfaces';
+import ComponentNode from './node/component';
 import { ServiceNode } from './node/service';
 import { TaskNode } from './node/task';
