+++ src/dependency-manager/src/graph/node/component.ts
